backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.5.2
pypdf==5.1.0
python-docx==1.1.0
openai>=1.66.0
tiktoken==0.8.0
scikit-learn==1.5.1
numpy==1.26.4
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
aiofiles==24.1.0
